Move test assertions from client pkg
