+++ test/api/application/delete_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/konveyor/tackle2-hub/api"
 	c "github.com/konveyor/tackle2-hub/test/api/client"
+	"github.com/konveyor/tackle2-hub/test/assert"
 )
 
 func TestApplicationDelete(t *testing.T) {
@@ -12,7 +13,7 @@ func TestApplicationDelete(t *testing.T) {
 	for _, r := range samples {
 		t.Run(r.Name, func(t *testing.T) {
 			// Create.
-			c.Must(t, Create(&r))
+			assert.Must(t, Create(&r))
 			rPath := c.Path(api.ApplicationRoot, c.Params{api.ID: r.ID})
 
 			// Try delete.
