+++ test/api/application/get_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/konveyor/tackle2-hub/api"
 	c "github.com/konveyor/tackle2-hub/test/api/client"
+	"github.com/konveyor/tackle2-hub/test/assert"
 )
 
 func TestApplicationGet(t *testing.T) {
@@ -12,7 +13,7 @@ func TestApplicationGet(t *testing.T) {
 	for _, r := range samples {
 		t.Run(r.Name, func(t *testing.T) {
 			// Create.
-			c.Must(t, Create(&r))
+			assert.Must(t, Create(&r))
 
 			// Try get.
 			got := api.Application{}
@@ -22,12 +23,12 @@ func TestApplicationGet(t *testing.T) {
 			}
 
 			// Assert the response.
-			if c.FlatEqual(got, samples) {
+			if assert.FlatEqual(got, samples) {
 				t.Errorf("Different response error. Got %v, expected %v", got, samples)
 			}
 
 			// Clean.
-			c.Must(t, Delete(&r))
+			assert.Must(t, Delete(&r))
 		})
 	}
 }
