+++ test/api/application/list_test.go
@@ -4,14 +4,14 @@ import (
 	"testing"
 
 	"github.com/konveyor/tackle2-hub/api"
-	c "github.com/konveyor/tackle2-hub/test/api/client"
+	"github.com/konveyor/tackle2-hub/test/assert"
 )
 
 func TestApplicationList(t *testing.T) {
 	samples := Samples()
 	// Create.
 	for i := range samples {
-		c.Must(t, Create(&samples[i])) // Need modify parent array instead of single record created by the loop (to keep created IDs).
+		assert.Must(t, Create(&samples[i])) // Need modify parent array instead of single record created by the loop (to keep created IDs).
 	}
 
 	// Try list.
@@ -25,12 +25,12 @@ func TestApplicationList(t *testing.T) {
 	if len(got) != len(samples) {
 		t.Errorf("Wrong list length. Got %d, expected %d.", len(got), len(samples))
 	}
-	if c.FlatEqual(got, samples) {
+	if assert.FlatEqual(got, samples) {
 		t.Errorf("Different response error. Got %v, expected %v", got, samples)
 	}
 
 	// Clean.
 	for _, r := range samples {
-		c.Must(t, Delete(&r))
+		assert.Must(t, Delete(&r))
 	}
 }
