+++ test/api/stakeholdergroup/api_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/konveyor/tackle2-hub/api"
 	c "github.com/konveyor/tackle2-hub/test/api/client"
+	"github.com/konveyor/tackle2-hub/test/assert"
 )
 
 func TestStakeholderGroupCRUD(t *testing.T) {
@@ -25,7 +26,7 @@ func TestStakeholderGroupCRUD(t *testing.T) {
 			if err != nil {
 				t.Errorf(err.Error())
 			}
-			if c.FlatEqual(got, &r) {
+			if assert.FlatEqual(got, &r) {
 				t.Errorf("Different response error. Got %v, expected %v", got, r)
 			}
 
@@ -63,7 +64,7 @@ func TestStakeholderGroupCRUD(t *testing.T) {
 func TestStakeholderGroupList(t *testing.T) {
 	samples := Samples()
 	for i := range samples {
-		c.Must(t, Create(&samples[i]))
+		assert.Must(t, Create(&samples[i]))
 	}
 
 	got := []api.StakeholderGroup{}
@@ -71,12 +72,12 @@ func TestStakeholderGroupList(t *testing.T) {
 	if err != nil {
 		t.Errorf(err.Error())
 	}
-	if c.FlatEqual(got, &samples) {
+	if assert.FlatEqual(got, &samples) {
 		t.Errorf("Different response error. Got %v, expected %v", got, &samples)
 	}
 
 	for _, r := range samples {
-		c.Must(t, Delete(&r))
+		assert.Must(t, Delete(&r))
 	}
 
 }
