+++ test/api/tag/api_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/konveyor/tackle2-hub/api"
 	c "github.com/konveyor/tackle2-hub/test/api/client"
+	"github.com/konveyor/tackle2-hub/test/assert"
 )
 
 func TestTagCRUD(t *testing.T) {
@@ -25,7 +26,7 @@ func TestTagCRUD(t *testing.T) {
 			if err != nil {
 				t.Errorf(err.Error())
 			}
-			if c.FlatEqual(got, &r) {
+			if assert.FlatEqual(got, &r) {
 				t.Errorf("Different response error. Got %v, expected %v", got, r)
 			}
 
@@ -65,7 +66,7 @@ func TestTagList(t *testing.T) {
 	samples := Samples()
 
 	for i := range samples {
-		c.Must(t, Create(&samples[i]))
+		assert.Must(t, Create(&samples[i]))
 	}
 
 	got := []api.Tag{}
@@ -73,12 +74,12 @@ func TestTagList(t *testing.T) {
 	if err != nil {
 		t.Errorf(err.Error())
 	}
-	if c.FlatEqual(got, samples) {
+	if assert.FlatEqual(got, samples) {
 		t.Errorf("Different response error. Got %v, expected %v", got, samples)
 	}
 
 	for _, r := range samples {
-		c.Must(t, Delete(&r))
+		assert.Must(t, Delete(&r))
 	}
 }
