+++ test/api/task/api_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/konveyor/tackle2-hub/api"
 	c "github.com/konveyor/tackle2-hub/test/api/client"
+	"github.com/konveyor/tackle2-hub/test/assert"
 )
 
 func TestTaskCRUD(t *testing.T) {
@@ -25,7 +26,7 @@ func TestTaskCRUD(t *testing.T) {
 			if err != nil {
 				t.Errorf(err.Error())
 			}
-			if c.FlatEqual(got, &r) {
+			if assert.FlatEqual(got, &r) {
 				t.Errorf("Different response error. Got %v, expected %v", got, r)
 			}
 
@@ -67,7 +68,7 @@ func TestTaskList(t *testing.T) {
 	samples := Samples()
 
 	for i := range samples {
-		c.Must(t, Create(&samples[i]))
+		assert.Must(t, Create(&samples[i]))
 	}
 
 	got := []api.Task{}
@@ -75,11 +76,11 @@ func TestTaskList(t *testing.T) {
 	if err != nil {
 		t.Errorf(err.Error())
 	}
-	if c.FlatEqual(got, samples) {
+	if assert.FlatEqual(got, samples) {
 		t.Errorf("Different response error. Got %v, expected %v", got, samples)
 	}
 
 	for _, r := range samples {
-		c.Must(t, Delete(&r))
+		assert.Must(t, Delete(&r))
 	}
 }
